refactor: moved the `State` struct into `src/indented_display.rs`
